# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cadence API
  version: 0.1.0
  description: >
    Source of truth for the Go server, the TypeScript client and the MCP tools.
    Naming (docs/spec/08-resolutions.md R1): operationId = MCP tool = UI command = <entity>.<verb>; the entity is
    declared once per path in x-cadence.entity; verbs come from api/vocabulary.yaml; the HTTP shape fixes the verb.
    Mutations are commands: Idempotency-Key, If-Match on the revision, ?dryRun=true; they emit one outbox event.
    Long work returns 202 {jobId}; gated work returns 202 {approvalId}. Errors are application/problem+json
    (RFC 9457) whose type resolves to a help article. Operations with x-cadence.planned are named here so the
    naming contract covers them; the server answers 501 until their phase implements them.
servers:
  - url: /api
...
paths:
...
  /host-sessions:release:
    x-cadence: { entity: hostSessions }
    post:
      operationId: hostSessions.release
      summary: Give back every live session of a host that is shutting down, so the next host takes them at once
      description: >
        The host stops its agents first (their last turn is committed and reported), then releases. Each session
        loses its host and token, a turn in progress counts as interrupted (the next host tells the agent, and its
        pending permission requests are withdrawn as interrupted, not declined), messages the host took but never gave
        the agent go back to pending, and the Chat shows the session as reconnecting until a host claims it.
      tags: [host]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/HostRelease" } } }
      responses:
        "200":
          description: The sessions released
          content: { application/json: { schema: { $ref: "#/components/schemas/HostReleased" } } }
        default: { $ref: "#/components/responses/Problem" }

  # ---------------------------------------------------------------- worker protocol (exempt: tag worker, never MCP tools; phase 2 · wave 0)
...
  /worker-leases/{id}:release:
    x-cadence: { entity: workerLeases }
    parameters: [{ $ref: "#/components/parameters/LeaseId" }]
    post:
      operationId: workerLeases.release
      summary: Complete a lease with the step's outcome, its output artifacts and final metrics, or a typed error
      tags: [worker]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/StepOutcome" } } }
      responses:
        "204": { description: Recorded; the control plane advances the pipeline }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
...
    LeaseId:
      name: id
      in: path
      required: true
      description: Lease id (lse_…)
      schema: { type: string }
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, description: "https://cadence.local/help/errors/<slug> — the help article for this error" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        currentRev: { type: integer, description: Present on 412 — the revision to rebase on }
        errors:
          type: array
          description: Field-level problems on 422
          items: { $ref: "#/components/schemas/ProblemFieldError" }
    ProblemFieldError:
      type: object
      required: [path, message]
      properties: { path: { type: string }, message: { type: string } }
...
    HostRelease:
      type: object
      required: [hostId]
      additionalProperties: false
      properties:
        hostId: { type: string, minLength: 1, maxLength: 200 }
        messages: { type: array, maxItems: 1000, items: { type: string }, description: "Messages this host took but never gave its agent; they go back to pending for the next host" }
    HostReleased:
      type: object
      required: [released]
      properties:
        released: { type: array, items: { type: string }, description: The ids of the sessions released }
...
    ArtifactRef:
      type: object
      required: [hash, type]
      properties:
        hash: { type: string, pattern: "^b3:[0-9a-f]{64}$" }
        type: { type: string, description: "shar, dataset, mix, base_model, checkpoint, training-state, calibration, hypotheses, analysis, eval-report, deployable, text" }
        size: { type: integer, format: int64, minimum: 0 }
        meta: { type: object, description: "Neutral, self-describing metadata of the type (R42)" }
...
    StepOutcome:
      type: object
      required: [state]
      additionalProperties: false
      properties:
        state: { type: string, enum: [done, failed, cancelled] }
        error: { $ref: "#/components/schemas/StepError" }
        outputs: { type: object, additionalProperties: { $ref: "#/components/schemas/ArtifactRef" } }
        metrics: { type: object, additionalProperties: { type: number }, description: "Final values (val_wer, seconds_per_step, …)" }
    StepError:
      type: object
      required: [type, message]
      properties:
        type: { type: string, enum: [oom, step, lost, cancelled, input], description: "oom gets one automatic retry at 0.75× batch; lost = reaped after missed heartbeats" }
        message: { type: string, maxLength: 4000 }
        retryable: { type: boolean }

    # ---------------------------------------------------------------- worker, queue, telemetry, runtimes schemas (stream W) — add below
...